+++ lib/engine/src/artifact.rs
@@ -18,12 +18,12 @@ use wasmer_vm::{
     TrapHandler, VMSharedSignatureIndex, VMTrampoline,
 };
 
-/// An `Artifact` is the product that the `Engine`
-/// implementation produce and use.
+/// An `Artifact` is the product of compiling a WASM module with an `Engine`. Conceptually this can
+/// be thought of as an object file or shared library that is produced as a result of a traditional
+/// ahead-of-time compilation flow.
 ///
-/// The `Artifact` contains the compiled data for a given
-/// module as well as extra information needed to run the
-/// module at runtime, such as [`ModuleInfo`] and [`Features`].
+/// The `Artifact` contains the compiled data for a given module as well as extra information
+/// needed to instantiate (load) the module and prepare the code therein to run.
 pub trait Artifact: Send + Sync + Upcastable + MemoryUsage {
     /// Return a reference-counted pointer to the module
     fn module(&self) -> Arc<ModuleInfo>;
@@ -35,11 +35,6 @@ pub trait Artifact: Send + Sync + Upcastable + MemoryUsage {
     ///
     /// Note: this will return `None` if the module is already instantiated.
     fn module_mut(&mut self) -> Option<&mut ModuleInfo>;
-
-    /// Register thie `Artifact` stack frame information into the global scope.
-    ///
-    /// This is required to ensure that any traps can be properly symbolicated.
-    fn register_frame_info(&self);
 
     /// Returns the features for this Artifact
     fn features(&self) -> &Features;
@@ -56,24 +51,6 @@ pub trait Artifact: Send + Sync + Upcastable + MemoryUsage {
     /// Returns data initializers to pass to `InstanceHandle::initialize`
     fn data_initializers(&self) -> &[OwnedDataInitializer];
 
-    /// Returns the functions allocated in memory or this `Artifact`
-    /// ready to be run.
-    fn finished_functions(&self) -> &BoxedSlice<LocalFunctionIndex, FunctionBodyPtr>;
-
-    /// Returns the function call trampolines allocated in memory of this
-    /// `Artifact`, ready to be run.
-    fn finished_function_call_trampolines(&self) -> &BoxedSlice<SignatureIndex, VMTrampoline>;
-
-    /// Returns the dynamic function trampolines allocated in memory
-    /// of this `Artifact`, ready to be run.
-    fn finished_dynamic_function_trampolines(&self) -> &BoxedSlice<FunctionIndex, FunctionBodyPtr>;
-
-    /// Returns the associated VM signatures for this `Artifact`.
-    fn signatures(&self) -> &BoxedSlice<SignatureIndex, VMSharedSignatureIndex>;
-
-    /// Get the func data registry
-    fn func_data_registry(&self) -> &FuncDataRegistry;
-
     /// Serializes an artifact into bytes
     fn serialize(&self) -> Result<Vec<u8>, SerializeError>;
 
@@ -82,80 +59,6 @@ pub trait Artifact: Send + Sync + Upcastable + MemoryUsage {
         let serialized = self.serialize()?;
         fs::write(&path, serialized)?;
         Ok(())
-    }
-
-    /// Do preinstantiation logic that is executed before instantiating
-    fn preinstantiate(&self) -> Result<(), InstantiationError> {
-        Ok(())
-    }
-
-    /// Crate an `Instance` from this `Artifact`.
-    ///
-    /// # Safety
-    ///
-    /// See [`InstanceHandle::new`].
-    unsafe fn instantiate(
-        &self,
-        tunables: &dyn Tunables,
-        resolver: &dyn Resolver,
-        host_state: Box<dyn Any>,
-        config: InstanceConfig,
-    ) -> Result<InstanceHandle, InstantiationError> {
-        self.preinstantiate()?;
-        let module = self.module();
-        let (imports, import_function_envs) = {
-            let mut imports = resolve_imports(
-                &module,
-                resolver,
-                &self.finished_dynamic_function_trampolines(),
-                self.memory_styles(),
-                self.table_styles(),
-            )
-            .map_err(InstantiationError::Link)?;
-
-            // Get the `WasmerEnv::init_with_instance` function pointers and the pointers
-            // to the envs to call it on.
-            let import_function_envs = imports.get_imported_function_envs();
-
-            (imports, import_function_envs)
-        };
-
-        // Get pointers to where metadata about local memories should live in VM memory.
-        // Get pointers to where metadata about local tables should live in VM memory.
-
-        let (allocator, memory_definition_locations, table_definition_locations) =
-            InstanceAllocator::new(&*module);
-        let finished_memories = tunables
-            .create_memories(&module, self.memory_styles(), &memory_definition_locations)
-            .map_err(InstantiationError::Link)?
-            .into_boxed_slice();
-        let finished_tables = tunables
-            .create_tables(&module, self.table_styles(), &table_definition_locations)
-            .map_err(InstantiationError::Link)?
-            .into_boxed_slice();
-        let finished_globals = tunables
-            .create_globals(&module)
-            .map_err(InstantiationError::Link)?
-            .into_boxed_slice();
-
-        self.register_frame_info();
-
-        let handle = InstanceHandle::new(
-            allocator,
-            module,
-            self.finished_functions().clone(),
-            self.finished_function_call_trampolines().clone(),
-            finished_memories,
-            finished_tables,
-            finished_globals,
-            imports,
-            self.signatures().clone(),
-            host_state,
-            import_function_envs,
-            config,
-        )
-        .map_err(|trap| InstantiationError::Start(RuntimeError::from_trap(trap)))?;
-        Ok(handle)
     }
 
     /// Finishes the instantiation of a just created `InstanceHandle`.
